Firebase config information added
The configuration object is held in a json file that will not be hosted on github.
However the logic is there for the information given during the firebase setup to be brought in trivially.
As well the firebase utilities imports in index.html have been brought down to the minimum working state until further notice.
In the future Analytics will be brought in.